requirements.txt
aiogram==3.13.1
feedparser==6.0.11
httpx==0.27.2
trafilatura==1.7.0
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
pydantic==2.9.2
rapidfuzz==3.9.6
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.2
sentence-transformers==3.2.1
torch==2.4.1
orjson==3.10.7
tqdm==4.66.5
tenacity==8.2.3
streamlit==1.37.1
